fix type mismatches in min/max calls to make the compiler happy

+++ OwlPatches/TapTempo.hpp
@@ -55,7 +55,7 @@ public:
   {
     if(abs(speed-s) > 16){
       int64_t delta = (int64_t)limit*(speed-s)/2048;
-      limit = max(1, limit+delta);
+      limit = max((int64_t)1, limit+delta);
       speed = s;
     }
   }
